requirements.txt
streamlit
flask
requests
google-cloud-speech
PyPDF2
pdfplumber
apscheduler